transfer tool - fixed many problems but STILL NOT FULLY FUNCTIONING YET, format, dirty commit I know, but idk anymore, anything to make to tool works, it's vibecode so I messed up, don't blame me =))

+++ source/join_sim/source/crash/crash.py
@@ -1,27 +1,39 @@
-import pygetwindow as gw
+import os
+import subprocess
+import time
+
 import psutil
 import win32process
-import subprocess
-import os
+
 from source.join_sim.source.logs import logger as logs
-from source.join_sim.source.utility import local_player,windows,recon_utils
-import time
+from source.join_sim.source.utility import local_player, recon_utils, windows
 from source.launcher.constants import GAME_WINDOW_TITLE
 
 appid = "2399830"
+crash_process: psutil.Process | None = None
+
 
 def detect_crash():
-    titles = set(gw.getAllTitles())
-    for title in titles:
-        if title == "The UE-ShooterGame Game has crashed and will close" or title == "Crash!":
+    global crash_process
+    for proc in psutil.process_iter(attrs=["name", "exe"]):
+        if proc.info["name"] == "CrashReportClient.exe":
+            crash_process = proc
             return True
-        
+    return False
+
+
 def close_game():
     try:
-        _, pid = win32process.GetWindowThreadProcessId(windows.hwnd) 
-        process = psutil.Process(pid)
+        global crash_process
+        if crash_process:
+            print(f"terminating crash process")
+            crash_process.terminate()
+            crash_process = None
 
-        process.terminate()
+        _, pid = win32process.GetWindowThreadProcessId(windows.hwnd)
+        process = psutil.Process(pid)
+        if process:
+            process.terminate()
         logs.logger.critical(f"game with pid {pid} terminated")
     except psutil.NoSuchProcess:
         logs.logger.critical("process not found")
@@ -30,27 +42,34 @@ def close_game():
     except Exception as e:
         logs.logger.critical(f"error: {e}")
 
+
 def launch_game_with_steam():
     steam_path = local_player.path("steam.exe")
-    print(steam_path)
     if os.path.exists(steam_path):
-    
         subprocess.run([steam_path, f"steam://run/{appid}"])
         logs.logger.critical(f"launching game with appid {appid} via steam")
     else:
-        logs.logger.critical("steam exe not found at the expected location cannot relaunch game")
+        logs.logger.critical(
+            "steam exe not found at the expected location cannot relaunch game"
+        )
+
 
 def re_open_game():
     close_game()
     time.sleep(10)
     launch_game_with_steam()
-    recon_utils.template_sleep_no_bounds("join_last_session",0.7,60)
-    windows.hwnd = windows.find_window_by_title(GAME_WINDOW_TITLE) # new process ID as game as relaunced
-    
+    recon_utils.template_sleep_no_bounds("join_last_session", 0.7, 60)
+    windows.hwnd = windows.find_window_by_title(
+        GAME_WINDOW_TITLE
+    )  # new process ID as game as relaunced
+
+
 def crash_rejoin():
     if detect_crash():
         close_game()
         time.sleep(10)
         launch_game_with_steam()
-        recon_utils.template_sleep_no_bounds("join_last_session",0.7,60)
-        windows.hwnd = windows.find_window_by_title(GAME_WINDOW_TITLE) # new process ID as game as relaunced
+        recon_utils.template_sleep_no_bounds("join_last_session", 0.7, 60)
+        windows.hwnd = windows.find_window_by_title(
+            GAME_WINDOW_TITLE
+        )  # new process ID as game as relaunced
